🚑 Playlist: fix how time_before_play is casted

+++ src/dao/sql/playlist.ts
@@ -314,7 +314,7 @@ SELECT
    LEFT OUTER JOIN playlist_content AS plc_current_playing ON plc_current_playing.pk_id_plcontent = pl.fk_id_plcontent_playing AND plc_current_playing.fk_id_playlist = :currentPlaylist_id
    LEFT OUTER JOIN playlist_content AS plc_before ON plc_before.pos BETWEEN COALESCE(plc_current_playing.pos, 0) AND (plc.pos) AND plc_before.fk_id_playlist = :currentPlaylist_id 
    WHERE plc_before.fk_kid = k.pk_kid     
-  ) AS time_before_play,
+  )::bigint AS time_before_play,
   pc.flag_visible AS flag_visible,
   ak.repository as repository,
   array_agg(DISTINCT pc_pub.pk_id_plcontent) AS public_plc_id,
